requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
cryptography==41.0.7
bcrypt==4.1.2

# CORS and HTTP
httpx==0.26.0
python-dateutil==2.8.2

# Logging and Monitoring
loguru==0.7.2
sentry-sdk[fastapi]==1.39.2
prometheus-client==0.19.0

# Additional utilities
pytz==2023.3.post1

# Rate limiting
slowapi==0.1.9

# Two-Factor Authentication
pyotp==2.9.0
qrcode==7.4.2
pillow==10.2.0

# Payment Processing
stripe==7.8.0

# Content Delivery and Storage
boto3==1.34.11
botocore==1.34.11

# PDF Generation
reportlab==4.0.8

# JWT for content access tokens
pyjwt==2.8.0

# Email Campaign Dependencies
jinja2==3.1.3
markupsafe==2.1.4
premailer==3.10.0
beautifulsoup4==4.12.3
celery[redis]==5.3.4
redis==4.6.0